Changed the car and the track
Updated the car asset to move the centre of mass to where the engine would be in the car allowing it to oversteer slightly.
Changed the torque curve of the car allowing it to have more torquer at higher RPM which seems to make it turn better.
Added a slight increase in the FOV when accelerating making it look like you are moving faster.
Added posts for a finish line on the track and rebuilt the lighting.