Fixed AI car stopping on test track and added start functionality
Added validation to the Player car to not allow you to set off unless the countdown timer has finished.
Added validation to the AI car to not move unless the countdown timer has finished.
Updated the GameLevel Level blueprint to tell the car what breaking functionality to do.
Updated the TestLevel Level blueprint to tell the car what breaking functionality to do.
Added a variable in the GameMode which can be used by blueprints to tell the cars to move when the timer has finished.